WIP: minor test improvements
