+++ integer/src/mul/ntt/mod.rs
@@ -585,110 +585,4 @@ mod tests {
         add_signed_mul_impl(&mut c, Positive, &a, &b, &mut memory);
         assert_eq!(&c[..], &expected[..], "sparse operand mismatch");
     }
-
-    /// Compare NTT against toom-3 at various sizes to find the crossover.
-    ///
-    /// Run with:
-    /// ```sh
-    /// # Pure toom-3 (prevent internal NTT recursion):
-    /// DASHU_THRESHOLD_NTT=99999999 cargo test -p dashu-int --release \
-    ///   -- ntt::tests::crossover --ignored --nocapture
-    /// ```
-    ///
-    /// The output is a table showing word count, transform length N,
-    /// toom-3 time, NTT time, and speedup ratio.  Use it to recalibrate
-    /// [`THRESHOLD_NTT`] after algorithmic changes.
-    #[test]
-    #[ignore]
-    #[allow(clippy::let_underscore_must_use)]
-    fn crossover() {
-        use std::time::Instant;
-
-        let sizes: &[usize] = &[
-            5_000, 10_000, 20_000, 30_000, 40_000, 50_000, 60_000, 80_000, 100_000, 120_000,
-            131_000,
-        ];
-
-        println!(
-            "{:>10} {:>4} {:>8} {:>12} {:>12} {:>10}",
-            "words", "bp", "N", "toom-3(ms)", "ntt(ms)", "ratio"
-        );
-        println!("{}", "-".repeat(68));
-
-        for &n in sizes {
-            let a: Vec<Word> = (0..n)
-                .map(|i| (i as u64 + 1).wrapping_mul(0x9E3779B97F4A7C15))
-                .collect();
-            let b: Vec<Word> = (0..n)
-                .map(|i| (i as u64 + 1).wrapping_mul(0xC6A4A7935BD1E995))
-                .collect();
-            let mut c_toom = vec![0u64; 2 * n];
-            let mut c_ntt = vec![0u64; 2 * n];
-
-            let layout = memory_requirement_up_to(2 * n, n);
-            let warmup = 2;
-            let iters = 5;
-
-            // toom-3 (may use NTT internally depending on DASHU_THRESHOLD_NTT)
-            let t_toom = {
-                let mut best = f64::MAX;
-                for _ in 0..warmup {
-                    let mut alloc = crate::memory::MemoryAllocation::new(layout);
-                    let mut mem = alloc.memory();
-                    c_toom.fill(0);
-                    let _ =
-                        crate::mul::toom_3::add_signed_mul(&mut c_toom, Positive, &a, &b, &mut mem);
-                }
-                for _ in 0..iters {
-                    let mut alloc = crate::memory::MemoryAllocation::new(layout);
-                    let mut mem = alloc.memory();
-                    c_toom.fill(0);
-                    let start = Instant::now();
-                    let _ =
-                        crate::mul::toom_3::add_signed_mul(&mut c_toom, Positive, &a, &b, &mut mem);
-                    let elapsed = start.elapsed().as_secs_f64() * 1000.0;
-                    if elapsed < best {
-                        best = elapsed;
-                    }
-                }
-                best
-            };
-
-            // NTT (direct)
-            let t_ntt = {
-                let mut best = f64::MAX;
-                for _ in 0..warmup {
-                    let mut alloc = crate::memory::MemoryAllocation::new(layout);
-                    let mut mem = alloc.memory();
-                    c_ntt.fill(0);
-                    let _carry = add_signed_mul_impl(&mut c_ntt, Positive, &a, &b, &mut mem);
-                }
-                for _ in 0..iters {
-                    let mut alloc = crate::memory::MemoryAllocation::new(layout);
-                    let mut mem = alloc.memory();
-                    c_ntt.fill(0);
-                    let start = Instant::now();
-                    let _carry = add_signed_mul_impl(&mut c_ntt, Positive, &a, &b, &mut mem);
-                    let elapsed = start.elapsed().as_secs_f64() * 1000.0;
-                    if elapsed < best {
-                        best = elapsed;
-                    }
-                }
-                best
-            };
-
-            assert_eq!(&c_ntt[..], &c_toom[..], "mismatch at n={n}");
-
-            let (_b_pack, nn, _k_eff) = select_params(n, n);
-            println!(
-                "{:>10} {:>4} {:>8} {:>12.3} {:>12.3} {:>9.2}x",
-                n,
-                _b_pack,
-                nn,
-                t_toom,
-                t_ntt,
-                t_ntt / t_toom
-            );
-        }
-    }
 }
